# Copies references to data structures to avoid extra ARC traffic.