Note: this change (by itself) does not significantly reduce ARC traffic, but in concert
with `UnmanagedBuffer` (#41), we see a ~15x performance improvement in
the parallelFor benchmark.

Performance for `NonBlockingThreadPool: parallel for, one level`:
```
master (b5ab4b6)                                   336813.5 ns  ±  65.93 %   6526        
managed buffers (9e31153)                         5062117.0 ns  ±  16.01 %    280         
this change (1ae770d)                              210097.0 ns  ±  58.77 %   9347        
tmp branch with managed buffers + this change       22078.0 ns  ± 942.16 %   8510        
```

Note: the variability in timing increases substantially as performance improves. This is (I believe) due to the lower frequency of stealing.