@dabrahams 2020-06-18
@saeta You didn't address my suggestion to turn on tsan.  Building parallel stuff and not testing with tsan seems really super inadvisable.  At least file an issue?